ensure log_to_stdout is patched for startup
In rocketjob/rails_semantic_logger#106 the order of requiring semantic
logger extensions is changed as part of the change to add support of the
"disabled" attribute. This appears to have the side-effect of not
patching "log_to_stdout", in
rails_semantic_logger/extensions/rails/server.rb, early enough to enable
logging to STDOUT by default.

This commit reverts that one change.