Use finish_with_state inside RailsSemanticLogger::Rack::Logger
This ensures that the finish instrumenter is sent to the same
subscribers that the initial request was sent to.

This solves an issue where in a threaded environment the dynamic
subscriptions from ActionDispatch::ServerTiming would cause a mismatch
in the number of subscriptions to a topic, which would pop too many
values off of the thread-local stacks, leading to invalid events being
sent to subscribers which leads nil value for `time` in Event object
which throw exception.