Respect action_dispatch.log_rescued_responses
This is a new configuration option in Rails 7 which defaults to true.
When an app sets this to false, this method does not log anything at all
for rescued responses.

Because rails_semantic_logger monkey patches
ActionDispatch::DebugExceptions and overrides this method, this
configuration option does not have any effect. Let's correct that.

More context on this configuration option:

- PR which introduced it https://github.com/rails/rails/pull/42592
- Evidence that this implementation should work fine in Rails 7.1+
  https://github.com/rails/rails/blob/v7.1.0/actionpack/lib/action_dispatch/middleware/debug_exceptions.rb#L137

I also noticed that rails_semantic_logger is not respecting the
action_dispatch.debug_exception_log_level configuration, and it makes
its own decisions about what is an appropriate level. I updated this to
also follow that configuration.

This is also how it works in Rails 7.1+:
https://github.com/rails/rails/blob/v7.1.0/actionpack/lib/action_dispatch/middleware/debug_exceptions.rb#L154